moved timestep_in_us to AbstractVertex
