+++ pacman/model/graphs/abstract_vertex.py
@@ -13,10 +13,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from six import add_metaclass
+from spinn_utilities.abstract_base import (
+    abstractproperty, AbstractBase)
 from pacman.exceptions import PacmanConfigurationException
 from pacman.model.graphs.common import ConstrainedObject
 
 
+@add_metaclass(AbstractBase)
 class AbstractVertex(ConstrainedObject):
     """ A vertex in a graph.
     """
@@ -76,3 +80,35 @@ class AbstractVertex(ConstrainedObject):
             If there is an attempt to add the same vertex more than once
         """
         self._added_to_graph = True
+
+    @abstractproperty
+    def timestep_in_us(self):
+        """ The timestep of this vertex in us
+
+        Typically will default to the machine timestep
+
+        If the machine vertexes have different timestemps this method will\
+        return the lowest common multiple of these
+
+        :rtype: int
+        """
+
+    def simtime_in_us_to_timesteps(self, simtime_in_us):
+        """
+        Helper function to convert simtime in us to whole timestep
+
+        This function verfies that the simtime is a multile of the timestep.
+
+        :param simtime_in_us: a simulation time in us
+        :type simtime_in_us: int
+        :return: the exact number of timeteps covered by this simtime
+        :rtype: int
+        :raises ValueError: If the simtime is not a mutlple of the timestep
+        """
+        n_timesteps = simtime_in_us // self.timestep_in_us
+        check = n_timesteps * self.timestep_in_us
+        if check != simtime_in_us:
+            raise ValueError(
+                "The requested time {} is not a multiple of the timestep {}"
+                "".format(simtime_in_us, self.timestep_in_us))
+        return n_timesteps
