+++ uinit_test_objects/simple_test_vertex.py
@@ -16,6 +16,7 @@
 """ test vertex used in many unit tests
 """
 from spinn_utilities.overrides import overrides
+from pacman.model.graphs import AbstractVertex
 from pacman.model.graphs.application import ApplicationVertex
 from pacman.model.graphs.machine import SimpleMachineVertex
 from pacman.model.resources import (
@@ -92,6 +93,6 @@ class SimpleTestVertex(ApplicationVertex):
         return self._n_atoms
 
     @property
-    @overrides(ApplicationVertex.timestep_in_us)
+    @overrides(AbstractVertex.timestep_in_us)
     def timestep_in_us(self):
         return 1000
